// src/entity/strava.entity.d.ts
export type TActivityType = 'AlpineSki' | 'BackcountrySki' | 'Canoeing' | 'Crossfit' | 'EBikeRide' | 'Elliptical' | 'Golf' | 'Handcycle' | 'Hike' | 'IceSkate' | 'InlineSkate' | 'Kayaking' | 'Kitesurf' | 'NordicSki' | 'Ride' | 'RockClimbing' | 'RollerSki' | 'Rowing' | 'Run' | 'Sail' | 'Skateboard' | 'Snowboard' | 'Snowshoe' | 'Soccer' | 'StairStepper' | 'StandUpPaddling' | 'Surfing' | 'Swim' | 'Velomobile' | 'VirtualRide' | 'VirtualRun' | 'Walk' | 'WeightTraining' | 'Wheelchair' | 'Windsurf' | 'Workout' | 'Yoga';
// ...
export interface ISummaryGear {
    id: string;//string 	The gear's unique identifier.
    resource_state: number;//integer 	Resource state, indicates level of detail. Possible values: 2 -> "summary", 3 -> "detail"
    primary: boolean;//boolean 	Whether this gear's is the owner's default one.
    name: string;//string 	The gear's name.
    distance: number; //float 	 The distance logged with this gear. 
}
// ...
export interface ISummaryClub {
    id: number;//long 	The club's unique identifier.
    resource_state: number;//integer 	Resource state, indicates level of detail. Possible values: 1 -> "meta", 2 -> "summary", 3 -> "detail"
    name: number;//string 	The club's name.
    profile_medium: string; //string 	URL to a 60x60 pixel profile picture.
    cover_photo: string;//string 	URL to a ~1185x580 pixel cover photo.
    cover_photo_small: string;//string 	URL to a ~360x176 pixel cover photo.
    sport_type?: string;//string 	Deprecated. Prefer to use activity_types. May take one of the following values: cycling, running, triathlon, other
    activity_types: TActivityType;// ActivityType 	The activity types that count for a club. This takes precedence over sport_type.
    city: string;//string 	The club's city.
    state: string;//string 	The club's state or geographical region.
    country: string; //string 	The club's country.
    private: boolean;//boolean 	Whether the club is private.
    member_count: number;//integer 	The club's member count.
    featured: boolean;//boolean 	Whether the club is featured or not.
    verified: boolean;//boolean 	Whether the club is verified or not.
    url: string;//string 	The club's vanity URL. 
}
// ...
export interface IDetailedAthlete {
    id: number;//long 	The unique identifier of the athlete
    username: string | null;// not defined inside doc
    resource_state: number;//integer 	Resource state, indicates level of detail. Possible values: 1 -> "meta", 2 -> "summary", 3 -> "detail"
    firstname: string; //string 	The athlete's first name.
    lastname: string;//string 	The athlete's last name.
    bio: string | null;
    profile_medium: string;//string 	URL to a 62x62 pixel profile picture.
    profile: string; //string 	URL to a 124x124 pixel profile picture.
    city: string;//string 	The athlete's city.
    state: string; //string 	The athlete's state or geographical region.
    country: string | null;//string 	The athlete's country.
    sex: "M" | "F";// string 	The athlete's sex. May take one of the following values: M, F
    premium: boolean;//boolean 	Deprecated. Use summit field instead. Whether the athlete has any Summit subscription.
    summit: boolean; //boolean 	Whether the athlete has any Summit subscription.
    created_at: string;//DateTime 	The time at which the athlete was created.
    updated_at: string; //DateTime 	The time at which the athlete was last updated.
    badge_type_id: null | number;
    friend: any | null; // ?????
    follower: any | null; // ????
    follower_count?: number;//integer 	The athlete's follower count.
    friend_count?: number;//integer 	The athlete's friend count.
    measurement_preference?: string; //string 	The athlete's preferred unit system. May take one of the following values: feet, meters
    ftp?: number;//integer 	The athlete's FTP (Functional Threshold Power).
    weight: number;//float 	The athlete's weight.
    clubs?: ISummaryClub; //SummaryClub 	The athlete's clubs.
    bikes?: ISummaryGear;//SummaryGear 	The athlete's bikes.
    shoes?: ISummaryGear;// SummaryGear 	The athlete's shoes. 
}

export interface IDetailedClub {
    id: number;//long 	The club's unique identifier.
    resource_state: number;//integer 	Resource state, indicates level of detail. Possible values: 1 -> "meta", 2 -> "summary", 3 -> "detail"
    name: string;//string 	The club's name.
    profile_medium: string;//string 	URL to a 60x60 pixel profile picture.
    cover_photo: string;//string 	URL to a ~1185x580 pixel cover photo.
    cover_photo_small: string; //string 	URL to a ~360x176 pixel cover photo.
    sport_type: string; //string 	Deprecated. Prefer to use activity_types. May take one of the following values: cycling, running, triathlon, other
    activity_types: TActivityType; //ActivityType 	The activity types that count for a club. This takes precedence over sport_type.
    city: string; //string 	The club's city.
    state: string; //string 	The club's state or geographical region.
    country: string; //string 	The club's country.
    private: boolean;//boolean 	Whether the club is private.
    member_count: number;//integer 	The club's member count.
    featured: boolean; //boolean 	Whether the club is featured or not.
    verified: boolean;//boolean 	Whether the club is verified or not.
    url: string; //string 	The club's vanity URL.
    membership: string; //string 	The membership status of the logged-in athlete. May take one of the following values: member, pending
    admin: boolean; //boolean 	Whether the currently logged-in athlete is an administrator of this club.
    owner: boolean; //boolean 	Whether the currently logged-in athlete is the owner of this club.
    following_count: number;//integer 	The number of athletes in the club that the logged-in athlete follows. 
}